Fixed an issue with WorkUnitManager
Currently WorkUnitManager runs a thread that keeps polling the work unit queue for new work units and blocks if the queue is empty. Because take() is a blocking method, this may cause the thread to be blocked when all tasks are completed. Changed take() to poll() with a timeout.

Signed-off-by: Yinan Li <ynli@linkedin.com>

RB=376756
R=etl-infra-reviewers
A=stakiar