Include retry count in the json document describing a task state
Currently, the json document generated by the method TaskState.toJson() does contain failure exception information if the task has failed in the first attempt or any of the retries but it doesn't contain information about the retry count. If the task finally succeeds after some retries, the task state will be SUCCESSFUL, which seems confusing given the exception information. This commit added retry count to the json document.

Signed-off-by: Yinan Li <ynli@linkedin.com>

RB=424149
R=stakiar
A=mitu