Added a tool for state store migration
Due to the package renaming, the package names of TaskState and JobState changed. Because the current state store uses SequenceFiles that store the fully-qualified class names of the key and value classes in the metafile, Gobblin won't be able to read the state store using the new TaskState and JobState classes unless we migrate the state store to use the new classes. This commit adds a utility for doing the migration.

Signed-off-by: Yinan Li <liyinan926@gmail.com>