Assign a multi-task ID to each MultiWorkUnit
Currently MultiWorkUnits do not have a task ID assigned to them. When the MRJobLauncher serializes a WorkUnit/MultiWorkUnit to a file, it uses the task ID assigned to the WorkUnit/MultiWorkUnit as the file name. Because a MultiWorkUnit does not have a task ID assigned to it, the file it gets serialized into does not have a proper file name. This commit fixes this by giving a MultiWorkUnit a so-called multi-task ID.

Signed-off-by: Yinan Li <liyinan926@gmail.com>