Changed the Extractor interface to disallow reuse of data records
Currently Etractor.readRecord() takes an instance of the data record class, which gives the Etractor chances to reuse the same data record. Originally, the purpose of having this was for performance reason. However, extracted data records now go to the forks through a blocking record queue, which may hold multiple references to the same data records if the writer is slower than the extractor. Allowing data record reuse will lead to undesired or even unexpected behavior if a data record is mutated while the queue is still holding more than one references to it. This commit changed the Extractor interface to disallow data record reuse.

Signed-off-by: Yinan Li <liyinan926@gmail.com>