Removed generated files from source
This allows for local jars (potentially propreitary jars that cannot be thrown in a public repository) to be included in gobblin.
You need to set the environment variable GOBBLIN_LOCAL_JAR_LOCATION with the path to the folder where you are storing your local jars for this to work.

TODO: temporarily setting transitive to true in gobblin-core's build.gradle in order to build. However, I am unsure about the effects on gobblin as a whole.