Do not add JobState to WorkUnits until when the tasks start
Currently in the MR mode, there's an optimization that avoids adding the JobState to the WorkUnits too early to reduce the memory footprint on the client side as well as reduce the cost of serializing the WorkUnits. The JobState is added to the WorkUnits, though, in the mappers after the WorkUnits are deserialized. Currently the Yarn mode is not doing this optimization so this PR fixes it.

Signed-off-by: Yinan Li <liyinan926@gmail.com>