Publish changes to improve HDFS performance.
Publishing now happens in two stages.  The first is to group all publish commands (source and destination path) by the source filesystem, source path root, destination filesystem, and destination path root.  Then we walk over the groups.  If the source and destination filesystems are equivalent and the filesystem supports atomically renaming a folder, we rename the source path root to the destination path root.  Otherwise, we run a move for each publish command.

NOTE: each publish command has an associated CommitAction which indicates the number of publishes which must succeed before the WorkUnitState is set to COMPLETE.