Handle race conditions involving rename in LocalFileSystem. (#1408)
* Handle race conditions involving rename in LocalFileSystem.
One race condition is when a directory D1/a/b/c/d is renamed to D2/a/b/c/d after another thread created D2/a/b/c/d.
This resulted in D2/a/b/c/d/d. Handle this by using the Java renameTo() method to avoid the LocalFileSystem fallback of copy.
The copy fallback also results in a race condition when a file is opened for copy in one thread and the parent directory is createed in another thread. A file not found exception is encountered in this case. This case is handled by catching the exception and retrying a few times.