Gobblin metrics refactoring 2 (#1442)
A POC for refactoring of gobblin-metrics to decrease external dependencies. No significant code changes:

    Created a new library gobblin-metrics-libs/gobblin-metrics-base which contains the basic APIs/implementations for gobblin-metrics
    Moved some common testing classes to gobblin.metrics.testing
    Created a new library gobblin-modules/gobblin-metrics-influxdb which contains the InfluxDB reporter
    Created a new library gobblin-modules/gobblin-metrics-graphite which contains the Graphite reporter
    Created a new library gobblin-modules/gobblin-metrics-hadoop which contains the Hadoop counters reporter

The previous gobblin-metrics library now depends on all of the above for backwards compatibility reasons. It also contains GobblinMetrics and GobblinMetricsRegistry which have explicit dependency on the different reporters. Removing that dependency will require some code refactoring. The library has been moved to gobblin-metrics-libs/gobblin-metrics to decrease the number of top-level deirectories.