[GOBBLIN-324] Add the cluster working directory config
Currently, the appWorkDir value is passed to the GobblinClusterManager
constructor and the GobblinTaskRunner constructor.
It's used to determine where the state files will be stored.
The default launch scripts call the main methods which pass in a
hardcoded "null" value and the code will take a default value like
file:/Users/username/standalone_cluster/1

It's useful to specify this value via a configuration.

When the config is not specified, the behavior is the same as before.

Also add some examples in the sample config file.

Testing:

Add new unit tests.
Manually run the cluster with and without this config.