[GOBBLIN-328] Fix GobblinClusterKillTest failure. Path mismatch
Issue:
org.apache.gobblin.cluster.GobblinClusterKillTest
failed because it looks at the wrong output path
for the output files.

Cause:
It appears that the paths have changed.

Fix:
Update the path to match.

Also:
* Instead of disabling the tests. Move the tests
to the group
disabledOnTravis to continue to disable it on
Travis until we have
higher confidence with these tests.
This makes it easier to run these tests locally.

* Run the setup and teardown methods per method.
This adds about 10s
overhead per method.
The worker kill test failed without this change
when running with other
tests. I suspect it is related to some state
management issues.
 This change makes tests more self contained.

* Some minor code quality improvements.

Closes #2180 from HappyRay/fix-cluster-kill-tests