[GOBBLIN-336] Rename SingleHelixTask to SingleTask
The class has no reference to Helix. It's used in
the child process.

Plan to create a new class named SingleHelixTask
in the next PR
to represent the Helix task instance.

Testing:

Integration tests passed.

The class has no reference to Helix. It's used in
the child process.

Plan to create a new class named SingleHelixTask
in the next PR
to represent the Helix task instance.

Testing:

Integration tests passed.

Closes #2218 from HappyRay/rename-helix-single-
task-class