Fix FindBug Warning
Warning:

Multithreaded correctness Warnings
Code	Warning
IS	Inconsistent synchronization of org.apache.gobblin.cluster.GobblinTaskRunner.taskStateModelFactory; locked 60% of time
Details
IS2_INCONSISTENT_SYNC: Inconsistent synchronization
The fields of this class appear to be accessed inconsistently with respect to synchronization.  This bug report indicates that the bug pattern detector judged that

The class contains a mix of locked and unlocked accesses,

Fix:

Initialize the field in the constructor instead of a private method.