[GOBBLIN-794] fix JsonIntermedidateToAvroConverter for nested array/object use cases
This commit fix a defect in converting nested json
schema, where a array had objects in items, and
then
the object had another array, and then another
layer of objects.

Converting the lowest level of objects require a
namespace, but the array converter did not carry
the
namespace to its nested converter.

The fix is to add a namespace parameter to
ArrayConverter, and let is pass the namespace to
subsequent calls
to RecordConverter.

Closes #2661 from chris9692/gobblin_794_fix