[GOBBLIN-1301] Get Hadoop token based on Azkaban env variable
Previously we used Azkaban job type to determine the presence of
existing Hadoop token. This had a problem with new job types
that were not in the predefined list.

In the new implementation, we check the presence of the token
directly in the environment variable.

The author of the previous implementation (ibuenos) could not recall
the reason to use job types, and had no objections for this change.