[GOBBLIN-1341] Make cluster name resolution more robust
[GOBBLIN-1341] Make cluster name resolution more
robust

We extend cluster name conversion to handle names
with
and without url schemes in config. In practice,
scheme
and port numbers are irrelevant, as a single host
usually
maps to a single cluster.

Handle use cases with local path and invalid urls

Clarify the logic based on PR comments

Closes #3177 from aplex/cluster-names