[GOBBLIN-1339] Add cluster name to dataset descriptor
[GOBBLIN-1339] Add cluster name to dataset
descriptor

We use dataset descriptors to track lineage.
Previously, it
only included the platform name (hive,hdfs) and
path of the
dataset. As a result, we could not differentiate
the data copy
between multiple production clusters, as the
dataset descriptors
were the same for them. We add an optional cluster
name to
address that.

This change will be used for data copy audit
system.

Hive and file-based copy code is updated to
include cluster names.

Use full storage system url instead of just well-
known name

Closes #3178 from aplex/dataset-cluster