[GOBBLIN-1451] Log better error messages from mappers
Previously, the root cause of the problems were wrapped in several
container exceptions that added little value in troubleshooting and
confused the users. They frequently reported that the job failed with
"RetryException" without looking at the root cause of it.

With this change, we'll unwrap the exceptions from tasks and show
the possible root cause at the top of the error message.

We also shorten the boilerplate messages, so that real root cause
is shown to the users right away.