[GOBBLIN-1458] Add endpoint to trigger a flow in GaaS (#3297)
Previously the way to manually trigger an execution of a scheduled flow is to send a partial update that sets the flow's runImmediately property to true. This PR improves the API by adding a restli action that sends the same partial update. 

Also copied the implementation of partialUpdateFlowConfig to FlowConfigV2ResourceLocalHandler for testing convenience (normally when running the service GobblinServiceFlowConfigResourceHandler would be used, but for tests it is FlowConfigV2ResourceLocalHandler).