[GOBBLIN-1494]Fix flaky test on HiveSource (#3338)
HiveSourceTest returns 2 watermarks for testGetWorkunitsAfterWatermark instead of 1.
This is due to System.getCurrentTime.millis() returning a different value when comparing against the filesystem last updated filetime. To guard against this, change the test to incorporate the workunit within the watermark by assigning it an earlier creation time, thus implying that the file was modified so it needs to be ingested again.