[GOBBLIN-1491] Add option to mark up/down D2 servers for gobblin service on leadership change (#3333)
Currently, D2 servers are started on all hosts at the time of service start. This change will allow using D2 delayed start, then marking up/down the servers when a host becomes the master/slave respectively. This means only the master will be announced to D2, so all requests will be routed to it.

The logic for markup is linkedin specific so must be implemented in internal code.