[GOBBLIN-1503] Modify mysql validation query to fail if db is in read-only mode
If mysql db enters read-only mode due to a failover, the connections need to be reset before writes will work. Currently the validation query is "select 1", so it will not evict connections in read-only mode. This new query is taken from https://stackoverflow.com/questions/39552146/evicting-connections-to-a-read-only-node-in-a-cluster-from-the-connection-pool, and should fail the validation if the db is read-only, so connections will be retried.