[GOBBLIN-1545] Wrap RestLiServiceException inside a CreateKVResponse (#3395)
Throwing exception in the service logs may confuse users (or log parsing scripts) that there is some issue with the service that requires attention. So we should not throw exception when it is a bad request and no issue from the service side. In these cases, we should wrap the exception in a Restli response that the caller can open.