[GOBBLIN-1832] Emit warning instead of failing job for retention of Hive Table Views (#3695)
We should not allow hive retention on a view since it shouldn't have access to delete underlying data. Instead it should throw a warning message if it is a view instead of failing the job as there may be retention jobs configured to include both hive tables and views. We want to be able to dynamically determine at runtime whether or not to skip retention on the dataset in question rather than statically allow/denylist tables in the configurations.

Co-authored-by: Urmi Mustafi <umustafi@linkedin.com>